ref: Use SWR in UserTabMedals;
