+++ app/user/[id]/components/Tabs/UserTabMedals.tsx
@@ -1,15 +1,13 @@
 import PrettyHeader from "@/components/General/PrettyHeader";
 import RoundedContent from "@/components/General/RoundedContent";
 import { Tooltip } from "@/components/Tooltip";
-import { getUserMedals } from "@/lib/actions/getUserMedals";
 import { GameMode } from "@/lib/types/GameMode";
-import { User } from "@/lib/types/User";
-import { UserMedal, UserMedals } from "@/lib/types/UserMedals";
 import { LucideMedal } from "lucide-react";
-import { useEffect, useState } from "react";
 import Image from "next/image";
 import PrettyDate from "@/components/General/PrettyDate";
 import Spinner from "@/components/Spinner";
+import { User, UserMedal, UserMedals } from "@/lib/hooks/api/user/types";
+import { useUserMedals } from "@/lib/hooks/api/user/useUserMedals";
 
 interface UserTabMedalsProps {
   user: User;
@@ -24,76 +22,13 @@ const MEDALS_NAMES: Record<string, string> = {
 };
 
 export default function UserTabMedals({ user, gameMode }: UserTabMedalsProps) {
-  const [isLoading, setIsLoading] = useState(false);
-
-  const [medals, setMedals] = useState<UserMedals | null>(null);
-
-  useEffect(() => {
-    if (isLoading) return;
-
-    setIsLoading(true);
-
-    getUserMedals(user.user_id, gameMode).then((data) => {
-      if (data.error) {
-        setIsLoading(false);
-        console.error(data.error);
-        return;
-      }
-
-      setMedals(data.data!);
-      setIsLoading(false);
-    });
-  }, [gameMode]);
-
-  const medalToElement = (medal: UserMedal) => {
-    const isAchieved = medal.unlocked_at !== null;
-
-    return (
-      <div className="flex flex-col items-center">
-        <Tooltip
-          key={medal.id}
-          content={
-            <div className="flex flex-col items-center text-center min-w-52">
-              <h2 className="text-xl font-semibold mx-auto text-wrap mb-2">
-                {medal.name}
-              </h2>
-
-              <div className="text-sm text-center whitespace-normal max-w-52 mx-auto mb-2 text-gray-200">
-                {medal.description}
-              </div>
-
-              <p className="text-xs text-gray-300">
-                {isAchieved ? (
-                  <div className="flex items-center">
-                    <p>achieved on&nbsp;</p>
-                    <PrettyDate time={medal.unlocked_at!} withTime={false} />
-                  </div>
-                ) : (
-                  `Not achieved`
-                )}
-              </p>
-            </div>
-          }
-        >
-          <Image
-            src={`https://a.${process.env.NEXT_PUBLIC_SERVER_DOMAIN}/medals/client/${medal.id}@2x.png`}
-            alt={medal.name}
-            width={75}
-            height={75}
-            className={`rounded-full ${
-              isAchieved ? "filter-none" : "filter grayscale opacity-50"
-            } center`}
-          />
-        </Tooltip>
-      </div>
-    );
-  };
+  const userMedalsQuery = useUserMedals(user.user_id, gameMode);
+  const userMedals = userMedalsQuery.data;
 
   return (
     <div className="flex flex-col">
       <PrettyHeader text="Medals" icon={<LucideMedal />} />
 
-      {/* map each mels */}
       <RoundedContent className="min-h-0 h-fit max-h-none grid grid-cols-2 gap-4">
         {Object.keys(MEDALS_NAMES).map((category) => (
           <div key={category} className="">
@@ -105,9 +40,9 @@ export default function UserTabMedals({ user, gameMode }: UserTabMedalsProps) {
               </div>
             </div>
             <div className="p-4 rounded-b-lg grid grid-cols-4 gap-4 justify-center items-center">
-              {medals ? (
-                medals[category as keyof UserMedals].medals.map((medal) =>
-                  medalToElement(medal)
+              {userMedals ? (
+                userMedals[category as keyof UserMedals].medals.map((medal) =>
+                  MedalElement(medal)
                 )
               ) : (
                 <div className="mx-auto col-span-4">
@@ -118,6 +53,50 @@ export default function UserTabMedals({ user, gameMode }: UserTabMedalsProps) {
           </div>
         ))}
       </RoundedContent>
+    </div>
+  );
+}
+
+function MedalElement(medal: UserMedal) {
+  const isAchieved = medal.unlocked_at !== null;
+
+  return (
+    <div className="flex flex-col items-center">
+      <Tooltip
+        key={medal.id}
+        content={
+          <div className="flex flex-col items-center text-center min-w-52">
+            <h2 className="text-xl font-semibold mx-auto text-wrap mb-2">
+              {medal.name}
+            </h2>
+
+            <div className="text-sm text-center whitespace-normal max-w-52 mx-auto mb-2 text-gray-200">
+              {medal.description}
+            </div>
+
+            <p className="text-xs text-gray-300">
+              {isAchieved ? (
+                <div className="flex items-center">
+                  <p>achieved on&nbsp;</p>
+                  <PrettyDate time={medal.unlocked_at!} withTime={false} />
+                </div>
+              ) : (
+                `Not achieved`
+              )}
+            </p>
+          </div>
+        }
+      >
+        <Image
+          src={`https://a.${process.env.NEXT_PUBLIC_SERVER_DOMAIN}/medals/client/${medal.id}@2x.png`}
+          alt={medal.name}
+          width={75}
+          height={75}
+          className={`rounded-full ${
+            isAchieved ? "filter-none" : "filter grayscale opacity-50"
+          } center`}
+        />
+      </Tooltip>
     </div>
   );
 }
